refactor(protocol): use uint64 for timestamp

+++ packages/protocol/script/debug.s.sol
@@ -105,7 +105,7 @@ contract DebugGasUsage is Test, IERC721Receiver {
       author: user1,
       app: user2,
       channelId: channelId,
-      deadline: block.timestamp + 1 days,
+      deadline: uint64(block.timestamp + 1 days),
       parentId: bytes32(0)
     });
