+++ packages/protocol/test/SharedFeeHook.t.sol
@@ -373,7 +373,7 @@ contract SharedFeeHookTest is Test, IERC721Receiver {
       author: user1,
       app: user2,
       channelId: channelId,
-      deadline: block.timestamp + 1 days,
+      deadline: uint64(block.timestamp + 1 days),
       parentId: bytes32(0)
     });
