+++ packages/protocol/test/TokenGatedHook.t.sol
@@ -110,7 +110,7 @@ contract TokenGatedHookTest is Test, IERC721Receiver {
       author: user1,
       app: user2,
       channelId: channelId,
-      deadline: block.timestamp + 1 days,
+      deadline: uint64(block.timestamp + 1 days),
       parentId: bytes32(0)
     });
 
@@ -145,7 +145,7 @@ contract TokenGatedHookTest is Test, IERC721Receiver {
       author: user1,
       app: user2,
       channelId: channelId,
-      deadline: block.timestamp + 1 days,
+      deadline: uint64(block.timestamp + 1 days),
       parentId: bytes32(0)
     });
 
@@ -178,7 +178,7 @@ contract TokenGatedHookTest is Test, IERC721Receiver {
       author: user1,
       app: user2,
       channelId: channelId,
-      deadline: block.timestamp + 1 days,
+      deadline: uint64(block.timestamp + 1 days),
       parentId: bytes32(0)
     });
